Fix: ScriptableObject.Manage: set array field
Moved type-conversion funtions to /Utils
