+++ Editor/Scripts/Utils/TypeConversionUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2b9f73da81f6094ca15ba58216a3917
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
